Revert "Formulario CLIENTE OK Formulado Pedido Quase OK Formulado Material OK"
This reverts commit c64610c3adaf5ac4df9f5cdc2d57b91eff1be0b2.

+++ src/modelos/Forma.java
@@ -1,23 +1,22 @@
 package modelos;
 
-import java.math.BigDecimal;
-
 /**
  *
  * @author babar
  */
 public abstract class Forma {
     
-    protected BigDecimal area;
+    protected double area;
         
-    public abstract BigDecimal calcularArea();
+    public abstract double calcularArea();
 
-    public BigDecimal getArea() {
+    public double getArea() {
         return area;
     }
 
-    public void setArea(BigDecimal area) {
+    public void setArea(double area) {
         this.area = area;
     }
+    
     
 }
